+++ libcob/move.c
@@ -2340,9 +2340,9 @@ cob_put_pointer (void *val, void *mem)
 }
 
 char *
-cob_get_picx (void *cbl_data, int len, void *char_field, int num_chars)
+cob_get_picx (void *cbl_data, size_t len, void *char_field, size_t num_chars)
 {
-	int		i;
+	size_t	i;
 	cob_u8_t	*p = cbl_data;
 
 	for (i = len - 1; i >= 0 && (p[i] == ' ' || p[i] == 0); i--);
@@ -2363,9 +2363,9 @@ cob_get_picx (void *cbl_data, int len, void *char_field, int num_chars)
 }
 
 void
-cob_put_picx( void *cbl_data, int len, void *string)
+cob_put_picx( void *cbl_data, size_t len, void *string)
 {
-	int	i,j;
+	size_t	i,j;
 	cob_u8_t	*p = cbl_data;
 	j = strlen ((char*)string);
 	if (j > len) {
